Change associate functions between book and author
The sequelize functions that were used to establish a relationship
between books and their authors are now changed to reflect the n:m
relation. Since the relation is m:n a junction table is needed, hence
the usage of the wrote model in the through option in the belongsToMany
function in both book and author. With this, the wrote model is now
serving its intended purpose.