Fix major bug in changing cjs module to ESmodule
Prior to this patch, models/index.js was being imported without
finishing the imports of all the models in the directory. Changing this
file to an ESmodule required the use of dynamic imports that needed to
be called asynchronously.
Thanks to the somewhat recent (and rightfully needed) top-level await,
files that include models/index.js must now wait for the async function
that initializes all of the modules to finish.