Create main api route for all database queries
This essentially done by renaming the server/routes/Book.js file
to "api.js". In the future, all queries to the database will access the
/api route. This was done as per greg's request because everything he
wishes upon this repository shall be. Edits to the JoinTestController
were also added to work test sequelize's join features. Another minor
change is the way the controller is exported in
server/controllers/Book.js. This is in preparation of making this
controller more reuseable in the future when other db queries will be
made to the book table.