Add SELinux support and remove the db volume entries for docker
Fedora has SELinux enabled by default, this was causing problems when
trying to run the development environment using docker. Adding the :z
flag resolves the docker permission issues.

The database file wasn't showing up in the docker container, so there
were a lot of file not found errors when trying to access the
database. After a bit of tinkering, I found that removing the db
volume labels fixed it. I suspect the issue was being caused due to an
attempt to mount a non-existant db directory at project root, where
the database file used to be; the db directory in the server
directory was being "covered" by the newly created db directory.
Changing which db directory was being mounted into the docker
container created syntax errors, and I didn't couldn't be bothered
to figurue out how to resolve them.
The server and client directories get mounted
into the containers anyway, so having a seperate db volume isn't needed.