Introduce Tags model to server and client
A controller for the Tags model has been created along with a new
endpoint for the api to fetch tags from: api/tags. In terms of the
client, the _search.svelte component now fetches all tags when mounted
in preparation for the future feature of suggesting tags as the user
types into the tags input.