Create tag searching functionality
The JoinTestController now utilizes more of sequelize's functionalities
to find books that have select tags specified by the client. For this
controller to be usable, a new route has been added to the api file. In
terms of the front-end, the index.svelte page now conditionally chooses
between the normal books endpoint and the JoinTestController endpoint
according to the user's intent (expressed through the tag search
parameter). The _search.svelte file now has a more fleshed out tag
search feature that: suggests tags as the user types a tag name, keeps a
visible collection of chosen tags, which can be directly be edited by
the user, updates matching books as user mutates the interanl tagArray,
and adds a valid tag automatically once the user has finished typing the
name of their desired tag.