Make a feature to copy search results to clipboard
Added a div element in the index.svelte page for the user to be
able to copy a link that represents the current search results. This is
an oversight that deals with how search results will be changed without
refreshing the page (and thus, the new results will not be reflected
in the URL search bar of the browser).