Update tagSearch results without refreshing page
When the user mutates the tag array, a svelte reaction is triggered that
will cause the searchStatus store to change state to "not_ready". This
state will change back to "ready" after a cool-down of 1.5 seconds. The
index page will update the results of the current tag search as a
reaction to the changes to the store. This was done to allow the user to
enter tags in the tagSearch input field without constantly causing the
page to refresh, which would be annoying if the user wanted to quickly
type their wanted tags.