Create a new endpoint for retrieving book authors
A new file was created for retrieving a single book's authors instead
of using the pre-existing books controller to keep the purpose of each
controller straighforward and maintain an organized project structure.

To be accessible, the api's new endpoint requires an id parameter. This
parameter still needs to be sanitized...