Allow cross-querying between tags & authors
Since the layout of the authors association with books is identical
to that of tags with books, allowing for users to use author names
as queries on tag search should be expected. The endpoint for quering
books with select tags now has the option of also getting books with
select authors. The same endpoint function is used to deliver results
with the different possible combinations in order to avoid duplicate
code. (the combinations being: queries with no authors and no tags,
queries with both authors and tags, etc...)