Fix templating issues with CSS in [id].svelte
Prior to this edit, the entries at the bottom of the book's main gird
was having alignment inconsistencies. Said alignments were more obvious
on different screen sizes. I added new CSS rules that restrict the
height of the bottom cells to ensure they maintain the same height
regardless of which nested flexbox they belong to.