Create front-end querying between tags & authors
In order to allow the book/[id].svelte page to make the book's listed
tags and authors interactable, the tagArray present in _search.svelte
must be turned into a store that can be updated through the book page.
A big side-effect of this is distinguishing between tokens that are
for authors and tokens that are for tags in the _search.svelte
component. While this makes the tagArray (now renamed as filterArray)
more complex, seperating between the two is more inuitive when setting
authors and tags to their respective values in the $page.searchParams
property (which is ultimately used to query the backend's database).
Finally, to make the feature of updating the $page stores searchParams
property reactive to any component that alters the tagArray store , the
tagSearchReaction was transformed into a svelte reaction block as it
should have been in the first place.