Uncomment migrations & seeding in entrypoint.sh
Before this change, the entrypoint shell that is specified by the server's
Dockerfile did not take into account the instatiation of the database. This
includes running the migration files with the sequelize cli and seeding the
created tables by running the seeder files. The initial intentions of the
server's dockerfile are now fulfilled by writing the database to its specified
volume.