Fix book edit and update in production
When building for production, the code gets obfuscated.
This caueses the edit and add book functions to not run, because
the function references in formSubmissionobj are being checked
via a hardcoded name.

This commit checks which function is which by checking whether
the function references match, instead of checking them by name.