Move edit endpoint res code to bottom of function
res.WriteHead(...) used to be placed before the logic responsible for
manipulating the database with sequelize. With the error-prone logic
inside of the editBook endpoint function, it is possible that
the ORM will throw an error/crash the server.Moving the response logic
to the bottom is more logical is it will prevent any potential false
positives to the client.