Refresh book search after adding/updating a book
The same logic used for triggering reactivity for updating search results
when adding a search tag to the tag search bar has been added to the
client's addbook and editbook functions. This full refresh that queriest
the entire database should not be too costly on the server to deploy.
querying is already done on the server's side, meaning if the number
of books in the database gets too big, pagination can be easily added.
This, in turn, will make the number of results predictable and would
offset the performance load off the server (and the client)